@@ -2836,61 +2836,13 @@ namespace Microsoft.PowerShell
                 // Output any training suggestions
                 try
                 {
-                    List<FeedbackEntry> feedbacks = FeedbackHub.GetFeedback(_parent.Runspace);
-                    if (feedbacks is null || feedbacks.Count == 0)
+                    List<FeedbackResult> feedbacks = FeedbackHub.GetFeedback(_parent.Runspace);
+                    if (feedbacks is null || feedbacks.Count is 0)
                     {
                         return;
                     }
 
-                    // Feedback section starts with a new line.
-                    ui.WriteLine();
-
-                    const string Indentation = "  ";
-                    string nameStyle = PSStyle.Instance.Formatting.FeedbackProvider;
-                    string textStyle = PSStyle.Instance.Formatting.FeedbackText;
-                    string ansiReset = PSStyle.Instance.Reset;
-
-                    if (!ui.SupportsVirtualTerminal)
-                    {
-                        nameStyle = string.Empty;
-                        textStyle = string.Empty;
-                        ansiReset = string.Empty;
-                    }
-
-                    int count = 0;
-                    var output = new StringBuilder();
-
-                    foreach (FeedbackEntry entry in feedbacks)
-                    {
-                        if (count > 0)
-                        {
-                            output.AppendLine();
-                        }
-
-                        output.Append("Suggestion [")
-                            .Append(nameStyle)
-                            .Append(entry.Name)
-                            .Append(ansiReset)
-                            .AppendLine("]:")
-                            .Append(textStyle);
-
-                        string[] lines = entry.Text.Split('\n', StringSplitOptions.RemoveEmptyEntries);
-                        foreach (string line in lines)
-                        {
-                            output.Append(Indentation)
-                                .Append(line.AsSpan().TrimEnd())
-                                .AppendLine();
-                        }
-
-                        output.Append(ansiReset);
-                        ui.Write(output.ToString());
-
-                        count++;
-                        output.Clear();
-                    }
-
-                    // Feedback section ends with a new line.
-                    ui.WriteLine();
+                    HostUtilities.RenderFeedback(feedbacks, ui);
                 }
                 catch (Exception e)
                 {
