@@ -35,8 +35,8 @@ namespace System.Management.Automation.ComInterop
         /// <returns>True if operation was bound successfully; otherwise, false.</returns>
         public static bool TryBindGetMember(GetMemberBinder binder, DynamicMetaObject instance, out DynamicMetaObject result, bool delayInvocation)
         {
-            Requires.NotNull(binder, nameof(binder));
-            Requires.NotNull(instance, nameof(instance));
+            Requires.NotNull(binder);
+            Requires.NotNull(instance);
 
             if (TryGetMetaObject(ref instance))
             {
@@ -66,9 +66,9 @@ namespace System.Management.Automation.ComInterop
         /// <returns>True if operation was bound successfully; otherwise, false.</returns>
         public static bool TryBindSetMember(SetMemberBinder binder, DynamicMetaObject instance, DynamicMetaObject value, out DynamicMetaObject result)
         {
-            Requires.NotNull(binder, nameof(binder));
-            Requires.NotNull(instance, nameof(instance));
-            Requires.NotNull(value, nameof(value));
+            Requires.NotNull(binder);
+            Requires.NotNull(instance);
+            Requires.NotNull(value);
 
             if (TryGetMetaObject(ref instance))
             {
@@ -91,9 +91,9 @@ namespace System.Management.Automation.ComInterop
         /// <returns>True if operation was bound successfully; otherwise, false.</returns>
         public static bool TryBindInvoke(InvokeBinder binder, DynamicMetaObject instance, DynamicMetaObject[] args, out DynamicMetaObject result)
         {
-            Requires.NotNull(binder, nameof(binder));
-            Requires.NotNull(instance, nameof(instance));
-            Requires.NotNull(args, nameof(args));
+            Requires.NotNull(binder);
+            Requires.NotNull(instance);
+            Requires.NotNull(args);
 
             if (TryGetMetaObjectInvoke(ref instance))
             {
@@ -116,9 +116,9 @@ namespace System.Management.Automation.ComInterop
         /// <returns>True if operation was bound successfully; otherwise, false.</returns>
         public static bool TryBindInvokeMember(InvokeMemberBinder binder, bool isSetProperty, DynamicMetaObject instance, DynamicMetaObject[] args, out DynamicMetaObject result)
         {
-            Requires.NotNull(binder, nameof(binder));
-            Requires.NotNull(instance, nameof(instance));
-            Requires.NotNull(args, nameof(args));
+            Requires.NotNull(binder);
+            Requires.NotNull(instance);
+            Requires.NotNull(args);
 
             if (TryGetMetaObject(ref instance))
             {
@@ -158,9 +158,9 @@ namespace System.Management.Automation.ComInterop
         /// <returns>True if operation was bound successfully; otherwise, false.</returns>
         public static bool TryBindGetIndex(GetIndexBinder binder, DynamicMetaObject instance, DynamicMetaObject[] args, out DynamicMetaObject result)
         {
-            Requires.NotNull(binder, nameof(binder));
-            Requires.NotNull(instance, nameof(instance));
-            Requires.NotNull(args, nameof(args));
+            Requires.NotNull(binder);
+            Requires.NotNull(instance);
+            Requires.NotNull(args);
 
             if (TryGetMetaObjectInvoke(ref instance))
             {
@@ -183,10 +183,10 @@ namespace System.Management.Automation.ComInterop
         /// <returns>True if operation was bound successfully; otherwise, false.</returns>
         public static bool TryBindSetIndex(SetIndexBinder binder, DynamicMetaObject instance, DynamicMetaObject[] args, DynamicMetaObject value, out DynamicMetaObject result)
         {
-            Requires.NotNull(binder, nameof(binder));
-            Requires.NotNull(instance, nameof(instance));
-            Requires.NotNull(args, nameof(args));
-            Requires.NotNull(value, nameof(value));
+            Requires.NotNull(binder);
+            Requires.NotNull(instance);
+            Requires.NotNull(args);
+            Requires.NotNull(value);
 
             if (TryGetMetaObjectInvoke(ref instance))
             {
@@ -208,8 +208,8 @@ namespace System.Management.Automation.ComInterop
         /// <returns>True if operation was bound successfully; otherwise, false.</returns>
         public static bool TryConvert(ConvertBinder binder, DynamicMetaObject instance, out DynamicMetaObject result)
         {
-            Requires.NotNull(binder, nameof(binder));
-            Requires.NotNull(instance, nameof(instance));
+            Requires.NotNull(binder);
+            Requires.NotNull(instance);
 
             if (IsComObject(instance.Value))
             {
@@ -245,7 +245,7 @@ namespace System.Management.Automation.ComInterop
         /// <returns>The collection of member names.</returns>
         internal static IList<string> GetDynamicDataMemberNames(object value)
         {
-            Requires.NotNull(value, nameof(value));
+            Requires.NotNull(value);
             Requires.Condition(IsComObject(value), nameof(value));
 
             return ComObject.ObjectToComObject(value).GetMemberNames(true);
@@ -260,7 +260,7 @@ namespace System.Management.Automation.ComInterop
         /// <returns>The collection of pairs that represent data member's names and their data.</returns>
         internal static IList<KeyValuePair<string, object>> GetDynamicDataMembers(object value, IEnumerable<string> names)
         {
-            Requires.NotNull(value, nameof(value));
+            Requires.NotNull(value);
             Requires.Condition(IsComObject(value), nameof(value));
 
             return ComObject.ObjectToComObject(value).GetMembers(names);
