@@ -3,7 +3,6 @@
 
 #nullable enable
 
-using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
@@ -14,6 +13,91 @@ using System.Threading;
 
 namespace System.Management.Automation.Subsystem.Feedback
 {
+    /// <summary>
+    /// Layout for displaying the recommended actions.
+    /// </summary>
+    public enum FeedbackDisplayLayout
+    {
+        /// <summary>
+        /// Display one recommended action per row.
+        /// </summary>
+        Portrait,
+
+        /// <summary>
+        /// Display all recommended actions in the same row.
+        /// </summary>
+        Landscape,
+    }
+
+    /// <summary>
+    /// The class represents a feedback item generated by the feedback provider.
+    /// </summary>
+    public sealed class FeedbackItem
+    {
+        /// <summary>
+        /// Gets the description message about this feedback.
+        /// </summary>
+        public string Header { get; }
+
+        /// <summary>
+        /// Gets the footer message about this feedback.
+        /// </summary>
+        public string? Footer { get; }
+
+        /// <summary>
+        /// Gets the recommended actions -- command lines or even code snippets to run.
+        /// </summary>
+        public List<string>? RecommendedActions { get; }
+
+        /// <summary>
+        /// Gets the layout to use for displaying the recommended actions.
+        /// </summary>
+        public FeedbackDisplayLayout Layout { get; }
+
+        /// <summary>
+        /// Gets or sets the next feedback item, if there is one.
+        /// </summary>
+        public FeedbackItem? Next { get; set; }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="FeedbackItem"/> class.
+        /// </summary>
+        /// <param name="header">The description message (must be not null or empty).</param>
+        /// <param name="actions">The recommended actions to take (optional).</param>
+        public FeedbackItem(string header, List<string>? actions)
+            : this(header, actions, footer: null, FeedbackDisplayLayout.Portrait)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="FeedbackItem"/> class.
+        /// </summary>
+        /// <param name="header">The description message (must be not null or empty).</param>
+        /// <param name="actions">The recommended actions to take (optional).</param>
+        /// <param name="layout">The layout for displaying the actions.</param>
+        public FeedbackItem(string header, List<string>? actions, FeedbackDisplayLayout layout)
+            : this(header, actions, footer: null, layout)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="FeedbackItem"/> class.
+        /// </summary>
+        /// <param name="header">The description message (must be not null or empty).</param>
+        /// <param name="actions">The recommended actions to take (optional).</param>
+        /// <param name="footer">The footer message (optional).</param>
+        /// <param name="layout">The layout for displaying the actions.</param>
+        public FeedbackItem(string header, List<string>? actions, string? footer, FeedbackDisplayLayout layout)
+        {
+            ArgumentException.ThrowIfNullOrEmpty(header);
+
+            Header = header;
+            RecommendedActions = actions;
+            Footer = footer;
+            Layout = layout;
+        }
+    }
+
     /// <summary>
     /// Interface for implementing a feedback provider on command failures.
     /// </summary>
@@ -27,29 +111,29 @@ namespace System.Management.Automation.Subsystem.Feedback
         /// <summary>
         /// Gets feedback based on the given commandline and error record.
         /// </summary>
-        /// <returns></returns>
-        string? GetFeedback(string commandLine, ErrorRecord lastError, CancellationToken token);
+        /// <param name="commandLine">The command line that was just executed.</param>
+        /// <param name="lastError">The error that was triggerd by the command line.</param>
+        /// <param name="token">The cancellation token to cancel the operation.</param>
+        /// <returns>The feedback item.</returns>
+        FeedbackItem? GetFeedback(string commandLine, ErrorRecord lastError, CancellationToken token);
     }
 
     internal sealed class GeneralCommandErrorFeedback : IFeedbackProvider
     {
         private readonly Guid _guid;
-        private readonly object[] _args;
-        private ScriptBlock? _fuzzySb;
 
         internal GeneralCommandErrorFeedback()
         {
             _guid = new Guid("A3C6B07E-4A89-40C9-8BE6-2A9AAD2786A4");
-            _args = new object[1];
         }
 
         public Guid Id => _guid;
 
-        public string Name => "General";
+        public string Name => "general";
 
         public string Description => "The built-in general feedback source for command errors.";
 
-        public string? GetFeedback(string commandLine, ErrorRecord lastError, CancellationToken token)
+        public FeedbackItem? GetFeedback(string commandLine, ErrorRecord lastError, CancellationToken token)
         {
             var rsToUse = Runspace.DefaultRunspace;
             if (rsToUse is null)
@@ -72,31 +156,31 @@ namespace System.Management.Automation.Subsystem.Feedback
 
                 if (command is not null)
                 {
-                    return StringUtil.Format(
-                        SuggestionStrings.Suggestion_CommandExistsInCurrentDirectory,
-                        target,
-                        localTarget);
+                    return new FeedbackItem(
+                        StringUtil.Format(SuggestionStrings.Suggestion_CommandExistsInCurrentDirectory, target),
+                        new List<string> { localTarget });
                 }
 
                 // Check fuzzy matching command names.
                 if (ExperimentalFeature.IsEnabled("PSCommandNotFoundSuggestion"))
                 {
-                    _fuzzySb ??= ScriptBlock.CreateDelayParsedScriptBlock(@$"
-                        param([string] $target)
-                        $cmdNames = Get-Command $target -UseFuzzyMatching -FuzzyMinimumDistance 1 | Select-Object -First 5 -Unique -ExpandProperty Name
-                        if ($cmdNames) {{
-                            [string]::Join(', ', $cmdNames)
-                        }}
-                    ", isProductCode: true);
+                    var pwsh = PowerShell.Create(RunspaceMode.CurrentRunspace);
+                    var results = pwsh.AddCommand("Get-Command")
+                            .AddParameter("UseFuzzyMatching")
+                            .AddParameter("FuzzyMinimumDistance", 1)
+                            .AddParameter("Name", target)
+                        .AddCommand("Select-Object")
+                            .AddParameter("First", 5)
+                            .AddParameter("Unique")
+                            .AddParameter("ExpandProperty", "Name")
+                        .Invoke<string>();
 
-                    _args[0] = target;
-                    var result = _fuzzySb.InvokeReturnAsIs(_args);
-
-                    if (result is not null && result != AutomationNull.Value)
+                    if (results.Count > 0)
                     {
-                        return StringUtil.Format(
+                        return new FeedbackItem(
                             SuggestionStrings.Suggestion_CommandNotFound,
-                            result.ToString());
+                            new List<string>(results),
+                            FeedbackDisplayLayout.Landscape);
                     }
                 }
             }
@@ -108,7 +192,6 @@ namespace System.Management.Automation.Subsystem.Feedback
     internal sealed class UnixCommandNotFound : IFeedbackProvider, ICommandPredictor
     {
         private readonly Guid _guid;
-        private string? _notFoundFeedback;
         private List<string>? _candidates;
 
         internal UnixCommandNotFound()
@@ -146,10 +229,7 @@ namespace System.Management.Automation.Subsystem.Feedback
             }
         }
 
-        /// <summary>
-        /// Gets feedback based on the given commandline and error record.
-        /// </summary>
-        public string? GetFeedback(string commandLine, ErrorRecord lastError, CancellationToken token)
+        public FeedbackItem? GetFeedback(string commandLine, ErrorRecord lastError, CancellationToken token)
         {
             if (Platform.IsWindows || lastError.FullyQualifiedErrorId != "CommandNotFoundException")
             {
@@ -176,15 +256,44 @@ namespace System.Management.Automation.Subsystem.Feedback
                 startInfo.RedirectStandardOutput = true;
 
                 using var process = Process.Start(startInfo);
-                var stderr = process?.StandardError.ReadToEnd().Trim();
-
-                // The feedback contains recommended actions only if the output has multiple lines of text.
-                if (stderr?.IndexOf('\n') > 0)
+                if (process is not null)
                 {
-                    _notFoundFeedback = stderr;
+                    string? header = null;
+                    List<string>? actions = null;
 
-                    var stdout = process?.StandardOutput.ReadToEnd().Trim();
-                    return string.IsNullOrEmpty(stdout) ? stderr : $"{stderr}\n{stdout}";
+                    while (true)
+                    {
+                        string? line = process.StandardError.ReadLine();
+                        if (line is null)
+                        {
+                            break;
+                        }
+
+                        if (line == string.Empty)
+                        {
+                            continue;
+                        }
+
+                        if (line.StartsWith("sudo ", StringComparison.Ordinal))
+                        {
+                            actions ??= new List<string>();
+                            actions.Add(line.TrimEnd());
+                        }
+                        else if (actions is null)
+                        {
+                            header = line;
+                        }
+                    }
+
+                    if (actions is not null && header is not null)
+                    {
+                        _candidates = actions;
+
+                        var footer = process.StandardOutput.ReadToEnd().Trim();
+                        return string.IsNullOrEmpty(footer)
+                            ? new FeedbackItem(header, actions)
+                            : new FeedbackItem(header, actions, footer, FeedbackDisplayLayout.Portrait);
+                    }
                 }
             }
 
@@ -206,37 +315,6 @@ namespace System.Management.Automation.Subsystem.Feedback
 
         public SuggestionPackage GetSuggestion(PredictionClient client, PredictionContext context, CancellationToken cancellationToken)
         {
-            if (_candidates is null && _notFoundFeedback is not null)
-            {
-                var text = _notFoundFeedback.AsSpan();
-                // Set to null to avoid potential race condition.
-                _notFoundFeedback = null;
-
-                // This loop searches for candidate results with almost no allocation.
-                while (true)
-                {
-                    // The line is a candidate if it starts with "sudo ", such as "sudo apt install python3".
-                    // 'sudo' is a command name that remains the same, so this check should work for all locales.
-                    bool isCandidate = text.StartsWith("sudo ", StringComparison.Ordinal);
-                    int index = text.IndexOf('\n');
-                    if (isCandidate)
-                    {
-                        var line = index != -1 ? text.Slice(0, index) : text;
-                        _candidates ??= new List<string>();
-                        _candidates.Add(new string(line.TrimEnd()));
-                    }
-
-                    // Break out the loop if we are done with the last line.
-                    if (index == -1 || index == text.Length - 1)
-                    {
-                        break;
-                    }
-
-                    // Point to the rest of feedback text.
-                    text = text.Slice(index + 1);
-                }
-            }
-
             if (_candidates is not null)
             {
                 string input = context.InputAst.Extent.Text;
@@ -263,7 +341,6 @@ namespace System.Management.Automation.Subsystem.Feedback
         public void OnCommandLineAccepted(PredictionClient client, IReadOnlyList<string> history)
         {
             // Reset the candidate state.
-            _notFoundFeedback = null;
             _candidates = null;
         }
 
