@@ -530,10 +530,11 @@ namespace System.Management.Automation.Language
             // Get the value of the index and value and call the compiler
             var index = indexExpressionAst.Index.Accept(this);
             var target = indexExpressionAst.Target.Accept(this);
-            
-            ArgumentNullException.ThrowIfNull(index, nameof(indexExpressionAst));
 
-            ArgumentNullException.ThrowIfNull(target, nameof(indexExpressionAst));
+            if (index is null || target is null)
+            {
+                throw new ArgumentNullException(nameof(indexExpressionAst));
+            }
 
             return GetIndexedValueFromTarget(target, index);
         }
